Fall back to writing a regular file without `O_TMPFILE`
Sadly `overlayfs` at least doesn't yet support `O_TMPFILE`
so we need the fallback path.

And this can get hit in usual cases like a build pod where
`/tmp` is actually part of the overlayfs and not a proper
host path.

This actually reveals an API problem in that the prefix/suffix
for this file isn't configurable; will file that as a
tracker for an API break.